services/nasa-data-ingest/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
aiohttp==3.9.1
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
celery==5.3.4
rasterio==1.3.9
xarray==2023.11.0
netcdf4==1.6.5
numpy==1.25.2
pandas==2.1.4
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
requests==2.31.0
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
